refactor: split up integration test and e2e tests more cleanly
