+++ das_client/app/integration_test/e2e/e2e_test_app.dart
@@ -7,10 +7,10 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import '../app_test.dart';
-import '../e2e_test_di.dart';
 import '../util/e2e_test_utils.dart';
 import '../util/test_utils.dart';
 import 'e2e_authenticator_override_scope.dart';
+import 'e2e_test_di.dart';
 import 'e2e_warnapp_override_scope.dart';
 
 class E2ETestApp {
