+++ das_client/app/integration_test/integration/integration_test_authenticated_scope.dart
@@ -11,27 +11,27 @@ import 'package:external_links/component.dart';
 import 'package:formation/component.dart';
 import 'package:logging/logging.dart';
 import 'package:ru_indications/component.dart';
-import 'package:train_identification/component.dart';
 import 'package:sfera/component.dart';
+import 'package:train_identification/component.dart';
 
-import 'mock_customer_oriented_departure_repository.dart';
-import 'mock_external_links_repository.dart';
-import 'mock_formation_repository.dart';
-import 'mock_ru_feature_provider.dart';
-import 'mock_ru_indications_repository.dart';
-import 'mock_train_identification_repository.dart';
-import 'mock_sim_train_view_model.dart';
-import 'mock_warn_app_view_model.dart';
+import '../mocks/mock_customer_oriented_departure_repository.dart';
+import '../mocks/mock_external_links_repository.dart';
+import '../mocks/mock_formation_repository.dart';
+import '../mocks/mock_ru_feature_provider.dart';
+import '../mocks/mock_ru_indications_repository.dart';
+import '../mocks/mock_sim_train_view_model.dart';
+import '../mocks/mock_train_identification_repository.dart';
+import '../mocks/mock_warn_app_view_model.dart';
 
-final _log = Logger('MockAuthenticatedScope');
+final _log = Logger('IntegrationTestAuthenticatedScope');
 
-class MockAuthenticatedScope extends AuthenticatedScope {
+class IntegrationTestAuthenticatedScope extends AuthenticatedScope {
   @override
-  String get scopeName => 'MockAuthenticatedScope';
+  String get scopeName => 'IntegrationTestAuthenticatedScope';
 
   @override
   Future<void> push() async {
-    _log.fine('Pushing mock scope $scopeName');
+    _log.fine('Pushing integration test scope $scopeName');
     getIt.pushNewScope(scopeName: scopeName);
 
     getIt.registerAuthProvider();
