+++ das_client/app/integration_test/integration/integration_test_journey_scope.dart
@@ -1,0 +1,6 @@
+import 'package:app/di/scopes/journey_scope.dart';
+
+class IntegrationTestJourneyScope extends JourneyScope {
+  @override
+  String get scopeName => 'IntegrationTestJourneyScope';
+}
