+++ das_client/app/integration_test/test/app_expiration_test.dart
@@ -5,6 +5,7 @@ import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
 import 'package:settings/component.dart';
 
 import '../app_test.dart';
+import '../integration/integration_test_app.dart';
 import '../mocks/mock_settings_repository.dart';
 import '../util/test_utils.dart';
