+++ das_client/app/integration_test/test/journey_notification_test.dart
@@ -8,6 +8,7 @@ import 'package:flutter_test/flutter_test.dart';
 
 import '../app_test.dart';
 import '../auth/integrationtest_authenticator.dart';
+import '../integration/integration_test_app.dart';
 import '../mocks/mock_customer_oriented_departure_repository.dart';
 import '../mocks/mock_ru_feature_provider.dart';
 import '../util/test_utils.dart';
