+++ das_client/app/integration_test/test/journey_search_overlay_test.dart
@@ -13,6 +13,7 @@ import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
 import 'package:train_identification/component.dart';
 
 import '../app_test.dart';
+import '../integration/integration_test_app.dart';
 import '../mocks/mock_train_identification_repository.dart';
 import '../util/test_utils.dart';
