+++ das_client/app/integration_test/test/journey_table_additional_speed_restriction_test.dart
@@ -5,7 +5,7 @@ import 'package:app/widgets/table/das_table.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../app_test.dart';
+import '../integration/integration_test_app.dart';
 import '../util/test_utils.dart';
 
 void main() {
