+++ das_client/app/integration_test/test/journey_table_advised_speeds_test.dart
@@ -6,6 +6,7 @@ import 'package:app/widgets/stickyheader/sticky_header.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import '../app_test.dart';
+import '../integration/integration_test_app.dart';
 import '../util/test_utils.dart';
 
 void main() {
