+++ das_client/app/integration_test/test/journey_table_brake_series_test.dart
@@ -2,6 +2,7 @@ import 'package:app/pages/journey/journey_screen/widgets/journey_table.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import '../app_test.dart';
+import '../integration/integration_test_app.dart';
 import '../util/test_utils.dart';
 
 void main() {
