+++ das_client/app/integration_test/test/journey_table_collapsible_rows_test.dart
@@ -12,6 +12,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:sfera/component.dart';
 
 import '../app_test.dart';
+import '../integration/integration_test_app.dart';
 import '../mocks/mock_sim_train_view_model.dart';
 import '../util/test_utils.dart';
