+++ das_client/app/integration_test/test/journey_table_track_equipment_test.dart
@@ -3,7 +3,7 @@ import 'package:app/pages/journey/journey_screen/widgets/table/cells/track_equip
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../app_test.dart';
+import '../integration/integration_test_app.dart';
 import '../util/test_utils.dart';
 
 void main() {
