+++ das_client/app/integration_test/test/short_term_changes_test.dart
@@ -6,7 +6,7 @@ import 'package:app/widgets/general_short_term_change_indicator.dart';
 import 'package:app/widgets/u_turn_indicator.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../app_test.dart';
+import '../integration/integration_test_app.dart';
 import '../util/test_utils.dart';
 
 void main() {
