+++ das_client/app/integration_test/test/suspicious_segment_test.dart
@@ -2,7 +2,7 @@ import 'package:app/pages/journey/journey_screen/notification/widgets/suspicious
 import 'package:flutter/foundation.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../app_test.dart';
+import '../integration/integration_test_app.dart';
 import '../util/test_utils.dart';
 
 void main() {
